+++ Cargo.toml
@@ -32,7 +32,7 @@ egui = { version = "0.15.0", default-features = false, features = ["persistence"
 glm = { version = "0.15.0", package = "nalgebra-glm" }
 image = { version = "0.23.14", default-features = false, features = ["png"] }
 device_query = "0.2.8"
-copypasta = "0.7.1"
+copypasta = {version = "0.7.1", default-features = false, features = ["x11"]}
 log = "0.4.14"
 simplelog = "0.10.1"
 fastrand = "1.5.0"
