+++ src/client/gui/mod.rs
@@ -1,1 +1,2 @@
 pub mod main_window;
+pub mod debug;
