+++ src/core/input/mod.rs
@@ -2,12 +2,13 @@ use std::{rc::Rc, sync::mpsc::Receiver};
 
 use glfw::{Glfw, WindowEvent};
 
-use crate::core::{input::glfw_input::GlfwInput, window::glfw_window::OverlayWindow};
+use crate::core::{input::glfw_input::GlfwInput, window::OverlayWindow};
 
 pub mod dq_input;
 pub mod glfw_input;
 pub mod rdev_input;
 
+#[derive(Debug)]
 pub struct InputManager {
     pub glfw_input: GlfwInput,
 }
@@ -27,6 +28,7 @@ impl InputManager {
 
 #[derive(Debug, Clone)]
 pub struct FrameEvents {
+    pub average_frame_rate: usize,
     pub all_events: Vec<WindowEvent>,
     pub time: f64,
     pub cursor_position: egui::Pos2,
