--- src/server.rs
@@ -1,5 +1,0 @@
-use crate::core::JokoCore;
-
-pub struct JokoServer {
-    pub core: JokoCore,
-}
